renamed package exercise to scanner and initiated studies about if,else and switch-case
