+++ src/loiane/exercise/scanner/ExerciseEight.java
@@ -1,4 +1,4 @@
-package loiane.exercise;
+package loiane.exercise.scanner;
 
 import java.util.Scanner;
