+++ src/loiane/exercise/scanner/ExerciseFourteen.java
@@ -1,4 +1,4 @@
-package loiane.exercise;
+package loiane.exercise.scanner;
 
 import java.util.Scanner;
