+++ src/loiane/exercise/scanner/ExerciseOne.java
@@ -1,4 +1,4 @@
-package loiane.exercise;
+package loiane.exercise.scanner;
 
 public class ExerciseOne {
     public static void main(String[] args) {
